cores/mor1kx: use 'version' instead of deprecated 'branch' option
When the the version option was introduced the branch option is
no longer recognised and will be ignored. Even though there is
a fallback to master when there is no version option stated,
explicitly stating it makes it more clear for users what to modify
to get a certain version.