cores: add verilog_utils core
The idea behind the verilog_utils core is to gather convinience
functions in one place, that other cores can pull in and use.
As a start, it contains an implementation of clog2, for
tools that doesn't yet have proper support for $clog2
(most notable ISE < 14.1).
Also functions for find first/last one are included.